+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 
 * 更新依赖项cbor2为5.8.0以修复CVE-2025-68131
 
+## 新增
+
+* 新增字段TokenInfo.pos_index以便于获取对应的正数index便于后续处理
+* 新增警告类ConfigDataPathSyntaxWarning以提供更好的配置路径语法警告信息
+
+## 变更
+
+* 字段TokenInfo.current_token由传入变为自动推导
+* 现在路径语法解析器发现孤立的键元信息会警告"Isolate meta found"
+* 现在路径语法解析器对无效的转义序列有更好的提示信息
+
 ## 修复
 
 * 修正MutableIndexed错误的类型注解